% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcc_biclustering.R
\name{biTGCC}
\alias{biTGCC}
\title{Tree-Guided Convex Bi-Clustering}
\usage{
biTGCC(
  data,
  lambdaSeq,
  gammaSeq,
  bandwidth_s = NULL,
  bandwidth_f = NULL,
  useNorm = TRUE,
  depthThresh_s = 10,
  depthThresh_f = 10,
  probThresh = 0.1,
  maxIter = 500,
  threshold = 1e-05,
  isNaive = TRUE
)
}
\arguments{
\item{data}{matrix with samples as rows and features as columns.}

\item{lambdaSeq}{tuning parameters sequence for samples.}

\item{gammaSeq}{tuning parameters sequence for features.}

\item{bandwidth_s}{bandwidth_s for the Gaussian kernel between samples.}

\item{bandwidth_f}{bandwidth_f for the Gaussian kernel between features.}

\item{useNorm}{logical, whether to use normalized distance. Default is `TRUE`.}

\item{depthThresh_s}{depth threshold target for the tree structure of samples. Default is `10`..}

\item{depthThresh_f}{depth threshold target for the tree structure of features. Default is `10`..}

\item{probThresh}{probability threshold for edge weight. Default is `0.1`.}

\item{maxIter}{maximum iteration of the Dykstra algorithm.}

\item{threshold}{stop condition of the Dykstra algorithm.}

\item{isNaive}{use naive method in searching MST. Default is `TURE`.}
}
\description{
Perform the tree-guided convex biclustering.
}
