% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{clusterLabel}
\alias{clusterLabel}
\title{Obtain cluster labels}
\usage{
clusterLabel(tgccFit, numClusters = 2)
}
\arguments{
\item{tgccFit}{output from `tgcc`, `biTGCC`, and `spTGCC`.}

\item{numClusters}{the desired number of clusters.}
}
\value{
A vector indicates the estimated clusters of samples.
}
\description{
obtain the estimated cluster labels with desired number of clusters
}
