% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dendrogram.R
\name{makeDendrogram}
\alias{makeDendrogram}
\title{Plot the dendrogram}
\usage{
makeDendrogram(tgccFit, labels)
}
\arguments{
\item{tgccFit}{output from `tgcc`, `biTGCC`, and `spTGCC`.}

\item{labels}{the underlying ground truth cluster labels.}
}
\value{
A dendrogram of hierarchical clustering indicates how samples are merged into clusters.
}
\description{
Plot the dendrogram based on the TGCC result.
}
