% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_matrices.R
\name{obtain_matrices_bitgcc}
\alias{obtain_matrices_bitgcc}
\title{obtain matrices from biTGCC output}
\usage{
obtain_matrices_bitgcc(bitgcc.fit)
}
\arguments{
\item{bitgcc.fit}{the result from biTGCC result}
}
\description{
obtain matrices from biTGCC output
}
