% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_matrices.R
\name{obtain_matrices_sptgcc}
\alias{obtain_matrices_sptgcc}
\title{obtain matrices from spTGCC output}
\usage{
obtain_matrices_sptgcc(sptgcc.fit)
}
\arguments{
\item{sptgcc.fit}{the result from spTGCC result}
}
\description{
obtain matrices from spTGCC output
}
