% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_matrices.R
\name{obtain_matrices_tgcc}
\alias{obtain_matrices_tgcc}
\title{obtain matrices from tgCC output}
\usage{
obtain_matrices_tgcc(tgcc.fit)
}
\arguments{
\item{tgcc.fit}{the result from tgCC function}
}
\description{
obtain matrices from tgCC output
}
