% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCC.R
\name{spTGCC}
\alias{spTGCC}
\title{sparse tree-guided convex clustering}
\usage{
spTGCC(
  input,
  lamseq,
  gamseq,
  bandwidth = NULL,
  normalize = TRUE,
  depth_thres = 50,
  prob = 0.1,
  max_iter = 500,
  precision_thres = 1e-05
)
}
\arguments{
\item{input}{matrix with samples as rows and features as columns.}

\item{lamseq}{tuning parameters sequence for rows.}

\item{bandwidth}{bandwidth for the Gaussian kernel between rows.}

\item{normalize}{to use normalized distance if true.}

\item{depth_thres}{threshold target of rows with in depth_thres.}

\item{prob}{threshold of edge weight.}

\item{max_iter}{maximum iteration of Dykstra algorithm.}

\item{precision_thres}{the stop condition of Dykstra algorithm.}
}
\description{
sparse tree-guided convex clustering
}
