% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcc-package.R
\docType{package}
\name{tgcc-package}
\alias{tgcc}
\alias{tgcc-package}
\title{tgcc: Tree Guided Convex Clustering}
\description{
This Package includes a dynamic programming based optimization framework of convex clustering.
}
\keyword{internal}
