% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgCC.R
\name{tgCC}
\alias{tgCC}
\title{Tree-Guided Convex Clustering}
\usage{
tgCC(
  data,
  lambdaSeq,
  bandwidth = NULL,
  useNorm = TRUE,
  depthThresh = 10,
  probThresh = 0.1
)
}
\arguments{
\item{data}{matrix with samples as rows and features as columns.}

\item{lambdaSeq}{sequence of tuning parameters.}

\item{bandwidth}{bandwidth for the Gaussian kernel between rows.}

\item{useNorm}{logical, whether to use normalized distance. Default is TRUE.}

\item{depthThresh}{depth threshold target. Default is 10.}

\item{probThresh}{probability threshold for edge weight. Default is 0.1.}
}
\value{
A list containing the clustered results and timings.
}
\description{
Tree-Guided Convex Clustering
}
