% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biCC.R
\name{biTGCC}
\alias{biTGCC}
\title{bi-tree-guided convex clustering}
\usage{
biTGCC(
  input,
  lamseq,
  gamseq,
  bandwidth_s = NULL,
  bandwidth_f = NULL,
  normalize = TRUE,
  depth_thres_s = 10,
  depth_thres_f = 10,
  prob = 0.1,
  max_iter = 500,
  precision_thres = 1e-05
)
}
\arguments{
\item{input}{matrix with samples as rows and features as columns.}

\item{lamseq}{tuning parameters sequence for rows.}

\item{gamseq}{tuning parameters sequence for columns.}

\item{bandwidth_s}{bandwidth_s for the Gaussian kernel between rows.}

\item{bandwidth_f}{bandwidth_f for the Gaussian kernel between columns.}

\item{normalize}{to use normalized distance if true.}

\item{depth_thres_s}{threshold target of rows with in depth_thres_s.}

\item{depth_thres_f}{threshold target of columns with in depth_thres_f.}

\item{prob}{threshold of edge weight.}

\item{max_iter}{maximum iteration of Dykstra algorithm.}

\item{precision_thres}{the stop condition of Dykstra algorithm.}
}
\description{
bi-tree-guided convex clustering
}
