% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcc_sparseclustering.R
\name{spTGCC}
\alias{spTGCC}
\title{Sparse Tree-Guided Convex Clustering}
\usage{
spTGCC(
  data,
  lambdaSeq,
  gammaSeq,
  bandwidth = NULL,
  useNorm = TRUE,
  depthThresh = 10,
  probThresh = 0.1,
  maxIter = 500,
  threshold = 1e-05,
  isNaive = FALSE
)
}
\arguments{
\item{data}{matrix with samples as rows and features as columns.}

\item{lambdaSeq}{tuning parameters sequence for samples.}

\item{gammaSeq}{tuning parameters sequence for features.}

\item{bandwidth}{bandwidth for the Gaussian kernel between rows. Default is heuristic assigned.}

\item{useNorm}{logical, whether to use normalized distance. Default is `TRUE`.}

\item{depthThresh}{depth threshold target. Default is `10`.}

\item{probThresh}{probability threshold for edge weight. Default is `0.1`.}

\item{maxIter}{maximum iteration for the Dykstra algorithm.}

\item{threshold}{stop condition for the Dykstra algorithm.}

\item{isNaive}{use naive method in searching MST. Default is `FALSE`.}
}
\value{
A list containing the clustered results.
}
\description{
Perform the sparse tree-guided convex clustering.
}
