% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcc.R
\name{tgCC}
\alias{tgCC}
\title{Tree-Guided Convex Clustering}
\usage{
tgCC(
  data,
  lambdaSeq,
  bandwidth = NULL,
  useNorm = TRUE,
  depthThresh = 10,
  probThresh = 0.1,
  isNaive = TRUE
)
}
\arguments{
\item{data}{matrix with samples as rows and features as columns.}

\item{lambdaSeq}{tuning parameters sequence.}

\item{bandwidth}{bandwidth for the Gaussian kernel between rows. Default is heuristic assigned.}

\item{useNorm}{logical, whether to use normalized distance. Default is `TRUE`.}

\item{depthThresh}{depth threshold target. Default is `10`.}

\item{probThresh}{probability threshold for edge weight. Default is `0.1`.}

\item{isNaive}{use naive method in searching MST. Default is `TURE`.}
}
\value{
A list containing the clustered results.
}
\description{
Perform the tree-guided convex clustering using dynamic programming method.
}
\examples{

library(tgcc)
dl <- tgcc:::make.mixgaussian(n = 400)
data <- dl$data
label <- dl$label

lambdaSeq <- seq(1, 2000, length.out = 100)
tgccFit <- tgCC( data = data, lambdaSeq = lambdaSeq)
makeDendrogram(tgccFit, label)

}
