% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgCC.R
\name{tgCC}
\alias{tgCC}
\title{tree-guided convex clustering}
\usage{
tgCC(
  input,
  lamseq,
  bandwidth = NULL,
  normalize = TRUE,
  depth_thres = 10,
  prob = 0.1
)
}
\arguments{
\item{input}{matrix with samples as rows and features as columns.}

\item{lamseq}{tuning parameters sequence.}

\item{bandwidth}{bandwidth for the Gaussian kernel between rows.}

\item{normalize}{to use normalized distance if true.}

\item{depth_thres}{threshold target with in depth_thres.}

\item{prob}{threshold of edge weight.}
}
\description{
tree-guided convex clustering
}
